requirements.txt
pandas
pydantic
google-genai
tqdm
openai
fasttext-langdetect
matplotlib
numpy
python-dotenv
